requirements.txt
python-dotenv~=1.0.1
openai~=1.44.0
tqdm~=4.66.5